requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
sqlalchemy>=2.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
